requirements.txt
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.2.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0

